% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usfiredepts}
\alias{usfiredepts}
\title{US Fire Departments Data}
\format{
\tabular{ll}{
\bold{Variable} \tab \bold{Description} \cr
FDID                              \tab Fire Department ID (assigned by state agencies and therefore not unique)\cr
Fire.dept.name                    \tab Fire Department Name                                                    \cr
HQ.addr1                          \tab Headquarters Address Line 1                                             \cr
HQ.addr2                          \tab Headquarters Address Line 2                                             \cr
HQ.city                           \tab Headquarters City                                                       \cr
HQ.state                          \tab Headquarters State                                                      \cr
HQ.zip                            \tab Headquarters Zip                                                        \cr
Mail.addr1                        \tab Mailing Address Line 1                                                  \cr
Mail.addr2                        \tab Mailing Address Line 2                                                  \cr
Mail.PO.box                       \tab Mailing Address PO Box                                                  \cr
Mail.city                         \tab Mailing Address City                                                    \cr
Mail.state                        \tab Mailing Address State                                                   \cr
Mail.zip                          \tab Mailing Address Zipcode                                                 \cr
HQ.phone                          \tab Headquarters Phone Number                                               \cr
HQ.fax                            \tab Headquarters Fax Number                                                 \cr
County                            \tab County                                                                  \cr
Dept.Type                         \tab Department Type                                                         \cr
Organization.Type                 \tab Organization Type                                                       \cr
Website                           \tab Website                                                                 \cr
Number.Of.Stations                \tab Number of Stations                                                      \cr
Active.Firefighters.Career        \tab Number of Active Firefighters (Career)                                  \cr
Active.Firefighters.Volunteer     \tab Number of Active Firefighters (Volunteer)                               \cr
Active.Firefighters.Paid.per.Call \tab Number of Active Firefighters (Paid Per Call)                           \cr
Non.Firefighting.Civilian         \tab Non-Firefighting Staff (Civilian)                                       \cr
Non.Firefighting.Volunteer        \tab Non-Firefighting Staff (Volunteer)                                      \cr
Primary.agency.for.emergency.mgmt \tab TRUE/FALSE if the agency is the primary emergency management agency
}
}
\source{
\url{https://apps.usfa.fema.gov/registry/download}
}
\usage{
usfiredepts
}
\description{
List of Fire Departments from the US Fire Department Registry.
}
\examples{
data(usfiredepts)
}
\keyword{datasets}
